Improved the compiler-test code, so the module compiler can be used by other applications
